*rm: AccumuoLayerFormat
